Add SVD interpretability feature to TransformerLens (#311)
* Changed head detection demo to install correct TransformerLens branch

* Added list of demos to tutorial section

* Added bullet points and fixed alphabetical order

* Added initial SVD Interpreter class. Get OV matrix and get top singular vectors methods.

* Improved code, added tests, completed functionality

* Added init for SVDInterpreter

* Fixed circular import, fixed a test

* Committing current error to show

* Finally got all unit tests working! Time to test numerical instability on PR

* Fixed einsum issue

* FactoredMatrix type support

* Fixed w_out issue

* Added demo

* Added documentation to SVD Interpreter

* Fixed utils.py merge issue

* Achieved 100% test coverage

* Added SVD demo to tutorials section

* Fixed formatting

* Fixed test formatting

* Fixed Literal 3.7 compatibility

* Fixed import order

* Added typechecked decorator to get_singular_vectors

* Black test formatting

+++ transformer_lens/utils.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import inspect
+import json
 import re
 import shutil
 from typing import Any, Callable, Dict, List, Optional, Tuple, Type, Union, cast
@@ -13,15 +14,13 @@ import transformers
 from datasets.arrow_dataset import Dataset
 from datasets.load import load_dataset
 from huggingface_hub import hf_hub_download
+from jaxtyping import Float, Int
 from rich import print as rprint
 from transformers import AutoTokenizer
 
 from transformer_lens import FactoredMatrix
 
 CACHE_DIR = transformers.TRANSFORMERS_CACHE
-import json
-
-from jaxtyping import Float, Int
 
 
 def select_compatible_kwargs(
@@ -477,10 +476,10 @@ class Slice:
             ValueError: If the slice is not an integer and max_ctx is not specified.
         """
         if self.mode == "int":
-            return np.array([self.slice])
+            return np.array([self.slice], dtype=np.int64)
         if max_ctx is None:
             raise ValueError("max_ctx must be specified if slice is not an integer")
-        return np.arange(max_ctx)[self.slice]
+        return np.arange(max_ctx, dtype=np.int64)[self.slice]
 
     def __repr__(
         self,
